+++ src/Card/index.stories.tsx
@@ -10,7 +10,51 @@ export default {
 export const Default: Story<CardProps> = (args) => (
   <Card title="Default size card" {...args}>
     <p>Card content</p>
-    <p>Card content</p>
-    <p>Card content</p>
+  </Card>
+);
+
+export const InnerCard: Story<CardProps> = (args) => (
+  <Card type="inner" title="Inner Card title" extra="Extra content" {...args}>
+    Inner Card content
+  </Card>
+);
+
+export const TabsCard: Story<CardProps> = (args) => (
+  <Card
+    title="Card with Tabs"
+    extra="Extra"
+    tabList={[
+      {
+        key: 'tab1',
+        tab: 'tab1',
+      },
+      {
+        key: 'tab2',
+        tab: 'tab2',
+      },
+    ]}
+    {...args}
+  >
+    some content
+  </Card>
+);
+
+const gridStyle = {
+  width: '25%',
+};
+
+export const CardsGrid: Story<CardProps> = (args) => (
+  <Card title="Cards Grid" {...args} style={{ width: 400 }}>
+    <Card.Grid style={gridStyle}>Content</Card.Grid>
+    <Card.Grid hoverable={false} style={gridStyle}>
+      Content
+    </Card.Grid>
+    <Card.Grid style={gridStyle}>Content</Card.Grid>
+  </Card>
+);
+
+export const CardMeta: Story<CardProps> = (args) => (
+  <Card hoverable style={{ width: 240 }} cover="Some cover" {...args}>
+    <Card.Meta title="Europe Street beat" description="www.google.com" />
   </Card>
 );