+++ src/Input/index.stories.tsx
@@ -1,21 +1,32 @@
 import { Story } from '@storybook/react';
-import React from 'react';
+import React, { useState } from 'react';
 
 import { NumericIDInput, NumericIDInputProps } from './NumericIDInput';
+import { InputNumberProps, InputProps } from './common';
 
-import { Input, InputNumber, InputNumberProps, InputProps } from './index';
+import { Input, InputNumber } from './index';
 
 export default {
   title: 'Input',
-  component: Input,
   onChange: { action: 'clicked' },
 };
 
 export const Text: Story<InputProps> = (args) => <Input {...args} />;
 
-export const Number: Story<InputNumberProps> = (args) => (
-  <InputNumber min={1} max={10} defaultValue={3} {...args} />
-);
+export const Number: Story<InputNumberProps> = (args) => {
+  const [num, setNum] = useState<number>(0);
+  return (
+    <InputNumber
+      min={1}
+      max={10}
+      value={num}
+      onChange={(val) =>
+        setNum(typeof val !== 'number' ? parseInt(val, 10) : val)
+      }
+      {...args}
+    />
+  );
+};
 
 export const NumericID: Story<NumericIDInputProps> = (args) => (
   <NumericIDInput {...args} />