+++ src/Select/index.stories.tsx
@@ -1,5 +1,5 @@
 import { Story } from '@storybook/react';
-import React from 'react';
+import React, { ReactElement } from 'react';
 
 import { Select, SelectProps } from './index';
 
@@ -8,7 +8,7 @@ export default {
 };
 
 export const Default: Story<SelectProps<string>> = (args) => (
-  <Select defaultValue="lucy" style={{ width: 120 }} {...args}>
+  <Select<string> defaultValue="lucy" style={{ width: 120 }} {...args}>
     <Select.Option value="jack">Jack</Select.Option>
     <Select.Option value="lucy">Lucy</Select.Option>
     <Select.Option value="henry" disabled>
@@ -16,3 +16,34 @@ export const Default: Story<SelectProps<string>> = (args) => (
     </Select.Option>
   </Select>
 );
+
+export const Multiple: Story<SelectProps<string>> = (args) => (
+  <Default mode="multiple" {...args} />
+);
+
+export const Group: Story<SelectProps<string>> = (args) => (
+  <Select<string> defaultValue="lucy" style={{ width: 120 }} {...args}>
+    <Select.OptGroup label="Guys">
+      <Select.Option value="jack">Jack</Select.Option>
+      <Select.Option value="henry" disabled>
+        Henry
+      </Select.Option>
+    </Select.OptGroup>
+    <Select.OptGroup label="Gals">
+      <Select.Option value="lucy">Lucy</Select.Option>
+    </Select.OptGroup>
+  </Select>
+);
+
+export const CustomDropdown: Story<SelectProps<string>> = (args) => (
+  <Multiple dropdownRender={renderDropdown} {...args} />
+);
+
+function renderDropdown(menu: ReactElement) {
+  return (
+    <>
+      <span>Custom Dropdown</span>
+      {menu}
+    </>
+  );
+}